Fix over-allocation in FallibleVec::try_reserve
Previously, the argument (named `new_cap`) was unconditionally added to the
existing capacity and the storage was reallocated. This seems to be
inconsistent with existing usage (and the Vec::reserve interface), which
call reserve with the number of additional elements which are about to be
added to the collection, presumably to avoid multiple allocations.

This changes the name of the argument to `additional` and changes the logic
to allocate space for that many additional insertions, considering the
current capacity, and not allocating unnecessarily.

Also, add some additional tests and remove the superfluous `_test` suffix
since test functions are already tagged with `#[test]` attributes.